*: make Destroy() void
There isn't anything we can do about the errors returned by Destroy().
All we ever do is log them, so lets not return the error and log it in
the Destroy() functions instead. This also means we do not abort from
cleanup code just because one section failed. This prevents one
destruction failure from causing other resources to be leaked because we
returned from Destory() early.