cork: pass through gpg sockets
bind mount ~/.gnupg to the same directory in the sdk. Mounting it other
places causes gpg to ignore the /run/user/$uid/gnupg directory and
create temp directories instead, although it is unclear why.

Always mount the /run/user/$uid/gnupg directory to the same location in
the sdk becuase configuring gpg to use a different directory is no
longer as simple as using the GPG_AGENT_INFO environment variable.

Don't be clever when mounting SSH sockets either. Just mount them in the
same place as they were in the real root.