cork: refactor enter, add legacy, bind-gpg-agent
 - Add --legacy option (default false). --legacy=true specifies the old
   cork enter behavior should be used.
 - Add --bind-gpg-agent option (default true). This indicates whether
   /run/user/$uid/gnupg should be bind mounted into the sdk.
 - Deprecate --experimental.
 - Make --experimental the same as specifying both --legacy=false and
   --bind-gpg-agent=false.
 - Make CmdDir a non-required field in the enter struct

cork enter --legacy=false uses the multicall library to spawn another
process of cork with special arguements so specifying an alternative
entry point. Previously not many arguments were passed to this
entrypoint so passing them as a list of strings was fine. Now the
complexity of the arguments passed has grown, so a struct is passed
instead by marshalling/unmarshalling it to json. This protects against
accidently specifying the wrong order or accidently omitting an argument
as well as providing type checking on the arguments passed.