platform/api/packet/api: update for new packngo sync
Two field names in the `DeviceCreateRequest` struct changed name
(`HostName` -> `Hostname` & `IPXEScriptUrl` -> `IPXEScriptURL`) and
pagination was added to the `Devices.List` call. The actual
implementation of the pagination options seems to mirror the `godo`
library.

For the `Devices.List` call it's fetching 1000 devices per page (the
documented maximum on their API docs).