platform/aws: check for missing Description before deref
I'm trying to use `ore` and hit a `SIGSEGV` because we're expecting all
the snapshot import tasks to have a description. Unfortunately, that is
not the case on the account I'm targeting. Since we ourselves always add
a description, we can safely skip those tasks since they can't be one of
ours anyway.