platform/api/aws: add AmazonS3ReadOnlyAccess to Role
When running kola on a root account that does not own the S3 objects
that are being fetched for tests the AmazonS3ReadOnlyAccess policy needs
to be added to the role (and the corresponding Instance Profile).

This also adds a conditional retry on machine creation to avoid errors
caused by the eventual consistency nature of AWS in relation to the
Instance Profile creation and subsequent use in machine creation.