kola/tests/ignition: adapt for RHCOS
The current ignition user tests were failing on RHCOS because RHCOS
does not have a docker group by default like CL.  There was a
subsequent failure due to the UID difference for the core user and
creation of new users.

The original tests were modified to only run on CL.  New tests were
added for RHCOS to account for no docker group and UID difference.