build: Use chmod, not chown for objects
This will fix: https://github.com/cgwalters/fedora-coreos-config/issues/1

Due to the way we're using hardlink checkouts for our roots, `chown`ing
the objects means that any data derived from them via scripts will
end up with that UID.

The *canonical* permissions in a `bare-user` OSTree repo are stored
as xattrs, but for new objects we won't know to use root.
(Probably we should require new objects unknown to a package to be
 root?)

All we really want here is for the objects to be accessible to
`pull-local`, so let's `chmod` and not `chown`.