main: Categorize commands
Promote the ones that operate on a "builddir" first.

`shell` is a bit special.

Then the other "utility" commands are like `virt-install` where
we expose it as its own function to make it easier for downstreams
to reuse the logic outside of `cmd-build` for more custom things.
We will need this until we implement a config file to define
which image types are built for example.