deps: Add virt-install bits, fix deps.txt
When I was originally splitting off `deps.txt` I ended up trying
to reorder/change things to use `/usr/bin` so I could test file
paths, but ended up backing that out.

In backing it out I apparently lost the dependency on the virt-install
stack.  Readd that.

And then that reveals the other problem, which is our dependencies
are a mix of file paths and packages, so support both of those.

(Potentially...we should build an rpm of our binaries
 and just use `Requires:`...)