cmds: Properly support --help
We had a bunch of copy-pasted stanzas that called out to `print_help`,
except that function is only defined in the `init` command.

Consistently define a `print_help` function, and make sure we support a
`-h/--help` option which prints it and exits nicely like in `init`.

Also explicitly check for too many args in functions that don't support
any extra positional args.