platform/local: increase interface limit beyond 8 bits
When all clusters are in the same netns, we'll eventually need more than
254 interfaces.  Support 16 bits' worth, but only create 500 for now,
since a large number of dhcp-host directives significantly increases
dnsmasq's startup time.

We don't actually need to retain the entire Interfaces array in memory
after dnsmasq.conf is rendered, but it's a small enough amount of RAM
not to be worth fixing for now.