Remove rpmdistro-gitoverlay
rdgo has some useful ideas, but lacks things like multi-arch
binding and being scalable for production builds.

We are going to move away from it for RHCOS, and as far as I know
the FCOS work does not use it.

The main reason to drop it though is it's a stumbling block for
building this container on other architectures.