Clean out /var for images
OSTree was designed from the very start to support having `/var`
be empty on boot, but for Fedora we ended up using Anaconda which
meant we had to carry a lot of hacks to run `systemd-tmpfiles`
in the root to support `%post`.  And when I say "hacks", I mean
seriously ugly hacks.

For Fedora CoreOS, since we're requiring Ignition and moving
away from Anaconda, we can use OSTree as it was designed
in a clean fashion.

For now since our build side is still using Anaconda, let's just
remove all contents in `/var` as a postprocessing step.

This is also a preparatory change for having images with a `/var`
partition pre-configured.

Requires: https://github.com/coreos/fedora-coreos-config/pull/29