libguestfish: Also mount `var` if such a partition exists
We discussed possibly using a `var` partition for FCOS in
https://github.com/coreos/fedora-coreos-tracker/issues/18

I would like to do so for my own Silverblue install, and possibly
for Silverblue by default.

So let's mount that partition if it exists, which means the other
code that cleans out what Anaconda did in `/var` will work.