cmdlib: Always rebuild supermin VM
So one major issue with caching the VM is that it doesn't carry over
well with updates of the container image: the `supermin.build/kernel`
symlink points to the `/lib/modules/<kver>/vmlinuz` file, which of
course will be broken as soon as the container is updated.

We could make caching smarter to handle this but... I don't know,
maintaining this cache invalidation logic doesn't seem worth it if
building it from scratch is not that much overhead.

This also makes e.g. hacking on ostree/rpm-ostree easier.