prune: Don't prune tagged builds
Follow-up to #243. Become aware of tags and make sure not to prune
builds that are tagged. I went the path of making tagged builds *not*
counting towards the `--keep=N` limit since that felt the most natural
to me.

One tricky bit is the OSTree repo pruning. The CLI is not really geared
towards this kind of pruning since we now could have older (tagged)
commits that we care about while some newer commits on that same branch
we don't care about. For now, I just swapped it to use
`--keep-younger-than`. Though this does align well with another
improvement I'd like to make to `prune` to learn timestamp-based pruning
instead.