compress: New command to compress all images
Add a new `compress` command that compresses all the images in a build.
I initially implemented this as e.g. `build --compress` and
`buildextend-* --compress`, but the `buildextend` commands need to
operate on the uncompressed qcow2.

Making it a separate command seems well-suited too though. It keeps the
other commands simpler and is mostly only useful for the production
pipeline case, where it's the penultimate step before uploading
artifacts.