cmd/prune: Add --keep-last-days=N switch
So right now, `prune` only knows `--keep=N` for fine-tuning the prune
operation. But humans are temporal beings and care more about the *age*
of a build rather than its index in the `builds[]` array.

This patch adds a `--keep-last-days=N` switch to accommodate this. I
initially was looking for a generic "fuzzy" parser that could parse e.g.
"2 weeks ago", and there is one (dateparser) but it's not packaged in
Fedora and I didn't want to start relying on pypi for this. Anyway, in
the prod context, I'd prefer we use something less fuzzy like
`--keep-last-days=14`.