convert python files to use `#!/usr/bin/env python3`
This will cause them to use the SCL on el7. It does make it
so that `python3 -u` on the shebang line doesn't work. We work
around this by setting PYTHONUNBUFFERED env var.

There is one file where we stick with using system python because
the SCL in EL7 doesn't have gobject library.