don't use --net=host
In dec465c we switched from using 9pfs to a network based solution
for sharing our ostree into the Anaconda install environment. As a
result of this we needed to predictably obtain the ip address of the
container and reliably know we could access an http server started
on the container's port 8000.

Using `--net=host` conflicts with this because it decreases the
probability of us choosing the right ip address and not requiring
us to poke holes in a firewall.