Revert "sdk: Regenerate the signature for the tests"
This reverts commit 39c84cbf7262af17609e495ab2fc8a19af7c532e.

The previous commit was to fix a misleading error message about
invalid base64 encoding, but it turns out the actual problem was
that ancient expired subkeys were dropped from the built-in keys.

We'll add back the expired subkeys so users can continue to operate
on ancient OS versions, and because of this, it is preferable to
keep testing that ancient signatures still work.