build: Drop FAHC, Ignition COPR
The FAHC stream was absolutely critical when coreos-assembler
was actively driving fixes/development in rpm-ostree.  Today
that isn't the case as much.  While I think we should
still ship git master rpm-ostree in cosa, we don't need to do
so right at this moment.  We can wait until we solve
https://github.com/coreos/fedora-coreos-tracker/issues/84

And we don't need the Ignition COPR right now.
(Although IMO it would make sense to add Ignition to FAHC but
 that's a distinct thing)

Closes: https://github.com/coreos/coreos-assembler/issues/336