add depcheck() to verify RPMs installed
When testing builds of the container image using different distros
(Fedora/EL7), it is possible for the image build to complete
successfully, only to find out that the `preflight()` check in
`src/cmdlib.sh` bails out because an RPM was not installed.

This change introduces `depcheck()` as a way to verify that the
requested RPMs were installed.  The new function is used in the
`build.sh` script and as part of `preflight()` in `src/cmdlib.sh`.