image-cloud.ks: Use DHCP_CLIENT_ID=mac
Right now in FCOS, VMs are getting *two* IP addresses when testing
locally in libvirt. The root of the issue is that the first request in
the initrd is using a different client ID than the one from NM in the
real root. This throws off at least dnsmasq, which then refuses to give
out the same IP again. Tell NM to also use the MAC address as the client
ID, to match what `dhclient` also does in the initrd.

For more background information, see:
https://github.com/coreos/fedora-coreos-config/issues/58

Closes: coreos/fedora-coreos-config#58