Include Anaconda in the container
It's part of the implementation detail of cosa.  Let's embed
it inside our container rather than downloading it dynamically.
This ensures that if you have the container you have *all* of the tools;
there's nothing separate to deal with and mirror.

It makes building the container more painful, but eh.