kola/tests/misc/selinux: refactor SELinux boolean tests
Previously, the SELinux boolean tests assumed a default value for the
boolean that was being tested.  When RHCOS flipped this value during
its compose, the tests broke.

This changes the tests to query the boolean first, then flip the value
of it during the test.

This introduces `toggleSelinuxBoolean()`, which accepts a SELinux
boolean to query and returns a struct with original + flipped values
to be used later.