kola/tests/rpmostree: use different pkg for 'rpm-ostree (un)install'
Move away from `fpaste` as the package to test with because it
requires Python2 on the host and that is not available in RHCOS.

I made the name of the package and the path to the binary variables
for easier reuse.  I also dropped the check on running the binary.

The tests pass the latest RHCOS and should pass FCOS, but I was
unable to run `kola` against FCOS.