Rebase tree to f30
This will allow us to match the content we're now composing. It also
means that from this point on, f29 is strictly about Fedora Atomic Host,
while f30 is strictly about Fedora CoreOS. This means more flexibility
wrt backwards compatibility.

The odd one is Anaconda; we're still using the f29 one for now. See
comment in patch for details.