src/cmd-archive: add command for archiving results to Koji.
Koji has the idea of a Content Generator [1], which allows external
processes to upload the build artifacts to a Koji Server. This command
will read in a build and upload supported files into Koji. The code was
tested against the internal Red Hat Koji instance with Red Hat CoreOS
builds.

[1] https://docs.pagure.org/koji/content_generators/