virt-install: Use tail --pid=$PID to ensure death
The `setpriv --pdeathsig` doesn't seem to be working on my system. Seems
like the parent `sh -c` does get cleaned up correctly, but not *its*
child `tail` process.

Use instead the built-in `--pid` flag which tells `tail` to stop
following and exit when we die. The `grep` naturally dies when `tail`
exits since its stdin is closed (and the parent `sh` soon follows).