mantle: bump to current master (9ef607a)
Adds support for Ignition v2s3 (as well as preserving spec2x support) to
mantle.

This should not be backported to the rhcos-4.1 branch unless other
external changes to the pipeline are done to pass `--ignition-version
v2` to `kola` as the default Ignition version that will be passed to
RHCOS machines is `v2s3`.