Add overrides/rootfs
This is strongly related to https://github.com/coreos/fedora-coreos-config/pull/66

For development iteration, we want the ability to directly overlay
binaries.  This depends on
https://github.com/projectatomic/rpm-ostree/pull/1830
which adds to rpm-ostree the ability to inject ostree refs into the build.

We could implement this in cosa by building an RPM (like we do for the overlay mentioned above)
but I think that's silly - we end up compressing the data only to immediately decompress it.
Plus, while there are advantages to having entries for files in the RPM
database, in practice we're lying about having these files built by RPM.
So let's just directly generate an OSTree commit from them and inject
that.

In practice for *this* use case rpm-ostree could actually learn
a way to directly consume a directory, but indirecting via OSTree
means it's equally easy to manage pre-generated ostree branches too
in an efficient way rather than re-committing each time.