build: Fix relative links for `src/config` with supermin
I use `ln -sr ~walters/src/redhat-coreos src/config` and the `-r`
makes a relative link which breaks this code as it tries to `mkdir -p`
on a relative path inside the supermin VM.

Rather than trying to make the path the same on the host and the VM,
add a `COSA_CONFIG_GIT` environment variable that overrides `src/config`,
and use that in the supermin path if we detect that `src/config` is
a symlink.  This way there doesn't need to be a fixed correspondence
between the host filesystem and VM (other than the rest of `workdir`
which should have all relative links contained inside it).